clean up use of types for x86_64
removed guesser_move_t, replaced with char
removed foe_move_t, replaced with size_t
removed index_t, replaced with size_t
replaced ad-hoc int and unsigned int usage with size_t

specialised makefile for core2 with march=core2

+++ evaluation.cpp
@@ -6,7 +6,7 @@ pair<eval_result_t, score_t> terminal_game_state(const context_t & ctx, const st
         return make_pair(EVAL_RESULT_BASE_LOSS, SCORE_GUESSER_LOSE);
     }
     // [BASE-WIN]: Alice wins
-    unsigned int n_words = h.live_word_indices.size();
+    size_t n_words = h.live_word_indices.size();
     assert(n_words > 0);
     if (n_words == 1) {
         return make_pair(EVAL_RESULT_BASE_WIN, SCORE_GUESSER_WIN);
@@ -19,18 +19,18 @@ pair<eval_result_t, score_t> terminal_game_state(const context_t & ctx, const st
     // least 1 possible word per guess, and she has at least
     // lives - 1 guesses left, and she wins if the number of
     // words is less than 2)
-    unsigned int lives = ctx.n_misses_for_loss - h.n_misses;
+    size_t lives = ctx.n_misses_for_loss - h.n_misses;
     if (h.live_word_indices.size() <= lives) {
         return make_pair(EVAL_RESULT_UPPER_BOUND_CHEAP, SCORE_GUESSER_WIN);
     }
 
     // needed for following two estimates
-    vector<unsigned int> unused_letter_indices = make_unused_letter_indices(
+    vector<size_t> unused_letter_indices = make_unused_letter_indices(
             ctx.letter_table, ctx.words.size(), h.guesses, h.live_word_indices);
 
     // only compute the expensive lower bound if a random forest classifier
     // predicts that it will succeed.
-    int forest_clf_inputs[6];
+    size_t forest_clf_inputs[6];
     forest_clf_inputs[0] = n_words;
     forest_clf_inputs[1] = lives;
     forest_clf_inputs[2] = h.guesses.size();
@@ -42,7 +42,7 @@ pair<eval_result_t, score_t> terminal_game_state(const context_t & ctx, const st
         // possible words left if the foe claims no subsequent guessed
         // letters are in the unknown word (this is able to a detect
         // guesser loss scenario early in some circumstances)
-        unsigned int lower_bound = lower_bound_on_remaining_words(ctx.letter_table,
+        size_t lower_bound = lower_bound_on_remaining_words(ctx.letter_table,
                 ctx.words.size(), h.live_word_indices, unused_letter_indices, lives);
         if (lower_bound > 1) {
             return make_pair(EVAL_RESULT_LOWER_BOUND_EXPENSIVE, SCORE_GUESSER_LOSE);
