+++ forest.h
@@ -1,59 +1,61 @@
 #ifndef FOREST_H
 #define FOREST_H
 
+#include <sys/types.h>
+
 /* decision tree classifier declarations */
-inline bool tree_clf_0(const int *x);
-inline bool tree_clf_1(const int *x);
-inline bool tree_clf_2(const int *x);
-inline bool tree_clf_3(const int *x);
-inline bool tree_clf_4(const int *x);
-inline bool tree_clf_5(const int *x);
-inline bool tree_clf_6(const int *x);
-inline bool tree_clf_7(const int *x);
-inline bool tree_clf_8(const int *x);
-inline bool tree_clf_9(const int *x);
-inline bool tree_clf_10(const int *x);
-inline bool tree_clf_11(const int *x);
-inline bool tree_clf_12(const int *x);
-inline bool tree_clf_13(const int *x);
-inline bool tree_clf_14(const int *x);
-inline bool tree_clf_15(const int *x);
-inline bool tree_clf_16(const int *x);
-inline bool tree_clf_17(const int *x);
-inline bool tree_clf_18(const int *x);
-inline bool tree_clf_19(const int *x);
-inline bool tree_clf_20(const int *x);
-inline bool tree_clf_21(const int *x);
-inline bool tree_clf_22(const int *x);
-inline bool tree_clf_23(const int *x);
-inline bool tree_clf_24(const int *x);
-inline bool tree_clf_25(const int *x);
-inline bool tree_clf_26(const int *x);
-inline bool tree_clf_27(const int *x);
-inline bool tree_clf_28(const int *x);
-inline bool tree_clf_29(const int *x);
-inline bool tree_clf_30(const int *x);
-inline bool tree_clf_31(const int *x);
-inline bool tree_clf_32(const int *x);
-inline bool tree_clf_33(const int *x);
-inline bool tree_clf_34(const int *x);
-inline bool tree_clf_35(const int *x);
-inline bool tree_clf_36(const int *x);
-inline bool tree_clf_37(const int *x);
-inline bool tree_clf_38(const int *x);
-inline bool tree_clf_39(const int *x);
-inline bool tree_clf_40(const int *x);
-inline bool tree_clf_41(const int *x);
-inline bool tree_clf_42(const int *x);
-inline bool tree_clf_43(const int *x);
-inline bool tree_clf_44(const int *x);
-inline bool tree_clf_45(const int *x);
-inline bool tree_clf_46(const int *x);
-inline bool tree_clf_47(const int *x);
-inline bool tree_clf_48(const int *x);
-inline bool tree_clf_49(const int *x);
+inline bool tree_clf_0(const size_t *x);
+inline bool tree_clf_1(const size_t *x);
+inline bool tree_clf_2(const size_t *x);
+inline bool tree_clf_3(const size_t *x);
+inline bool tree_clf_4(const size_t *x);
+inline bool tree_clf_5(const size_t *x);
+inline bool tree_clf_6(const size_t *x);
+inline bool tree_clf_7(const size_t *x);
+inline bool tree_clf_8(const size_t *x);
+inline bool tree_clf_9(const size_t *x);
+inline bool tree_clf_10(const size_t *x);
+inline bool tree_clf_11(const size_t *x);
+inline bool tree_clf_12(const size_t *x);
+inline bool tree_clf_13(const size_t *x);
+inline bool tree_clf_14(const size_t *x);
+inline bool tree_clf_15(const size_t *x);
+inline bool tree_clf_16(const size_t *x);
+inline bool tree_clf_17(const size_t *x);
+inline bool tree_clf_18(const size_t *x);
+inline bool tree_clf_19(const size_t *x);
+inline bool tree_clf_20(const size_t *x);
+inline bool tree_clf_21(const size_t *x);
+inline bool tree_clf_22(const size_t *x);
+inline bool tree_clf_23(const size_t *x);
+inline bool tree_clf_24(const size_t *x);
+inline bool tree_clf_25(const size_t *x);
+inline bool tree_clf_26(const size_t *x);
+inline bool tree_clf_27(const size_t *x);
+inline bool tree_clf_28(const size_t *x);
+inline bool tree_clf_29(const size_t *x);
+inline bool tree_clf_30(const size_t *x);
+inline bool tree_clf_31(const size_t *x);
+inline bool tree_clf_32(const size_t *x);
+inline bool tree_clf_33(const size_t *x);
+inline bool tree_clf_34(const size_t *x);
+inline bool tree_clf_35(const size_t *x);
+inline bool tree_clf_36(const size_t *x);
+inline bool tree_clf_37(const size_t *x);
+inline bool tree_clf_38(const size_t *x);
+inline bool tree_clf_39(const size_t *x);
+inline bool tree_clf_40(const size_t *x);
+inline bool tree_clf_41(const size_t *x);
+inline bool tree_clf_42(const size_t *x);
+inline bool tree_clf_43(const size_t *x);
+inline bool tree_clf_44(const size_t *x);
+inline bool tree_clf_45(const size_t *x);
+inline bool tree_clf_46(const size_t *x);
+inline bool tree_clf_47(const size_t *x);
+inline bool tree_clf_48(const size_t *x);
+inline bool tree_clf_49(const size_t *x);
 
 /* forest classifier */
-bool forest_clf(const int *x);
+bool forest_clf(const size_t *x);
 
 #endif /* FOREST_H */
